requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-writable-nested==0.5.1
llvmlite==0.30.0
Pillow==6.2.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
